requirements.txt
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.100.0
pyttsx3==2.90
moviepy==1.0.3
requests==2.31.0
